virtio/nvme: `num_of-vfs` field update
Changes made by 'PRM Rev 0.55 Draft A' about `num_of_vfs`:
1. Removed num_of_vfs and device_pci_parameters from VIRTO DEVICE
2. Removed num_of_vfs from NVME DEVICE EMULATION Object
3. Added num_of_vfs to DEVICE PCI PARAMETERS

+++ src/snap_nvme.h
@@ -120,7 +120,7 @@ struct snap_nvme_device_attr {
 	bool					enabled;
 	struct snap_nvme_registers		bar;
 	uint64_t				modifiable_fields;//mask of snap_nvme_device_modify
-	int					num_of_vfs;
+	uint16_t				num_of_vfs;
 	uint32_t				crossed_vhca_mkey;
 };
